Changed default download limit to unlimited (-1); cleaned up and formated code

+++ src/main/java/com/ftd/smartshare/server/SQLRequestHandler.java
@@ -18,21 +18,24 @@ public class SQLRequestHandler {
 	private static final String PASSWORD = "bondstone";
 
 	public SQLRequestHandler() {
-	}
-
-	public boolean connect() {
 		try {
 			Class.forName("org.postgresql.Driver");
-			return true;
 		} catch (ClassNotFoundException e) {
 			System.out.println("Unable to load the postgreSQL Driver.");
-			return false;
 		}
 	}
 
+	/**
+	 * Get File from SQL Database
+	 *
+	 * @param downloadRequestDto JAXB annotated class representing the download
+	 *                           request
+	 * @return return UploadRequestDto, return null if not found or expired
+	 */
 	public UploadRequestDto getFile(DownloadRequestDto downloadRequest) {
 		UploadRequestDto downloadedFile = null;
 		try (Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);) {
+			// Get file with the given file_name and password
 			String sqlReturn = "SELECT * FROM smartshare.files WHERE file_name = ? AND password = ? ;";
 			PreparedStatement preparedStatementReturn = connection.prepareStatement(sqlReturn);
 			preparedStatementReturn.setString(1, downloadRequest.getFilename());
@@ -47,19 +50,22 @@ public class SQLRequestHandler {
 					downloadedFile.setFilename(resultSet.getString("file_name"));
 					downloadedFile.setFile(resultSet.getBytes("file"));
 
-					// If max downloads reached (after this download) delete
-					if (resultSet.getInt("max_downloads") <= resultSet.getInt("total_downloads") + 1) {
-						PreparedStatement preparedStatement = connection
-								.prepareStatement("DELETE FROM smartshare.files WHERE id = ?;");
-						preparedStatement.setInt(1, resultSet.getInt("id"));
-						preparedStatement.execute();
-					} else { // If max downloads not reached
-						// Increment total_downloads
-						PreparedStatement preparedStatement = connection
-								.prepareStatement("UPDATE smartshare.files SET total_downloads = ? WHERE id = ?;");
-						preparedStatement.setInt(1, (resultSet.getInt("total_downloads") + 1));
-						preparedStatement.setInt(2, resultSet.getInt("id"));
-						preparedStatement.execute();
+					// -1 means the file has unlimited downloads
+					if (resultSet.getInt("max_downloads") != -1) {
+						// If max downloads reached (after this download) delete
+						if (resultSet.getInt("max_downloads") <= resultSet.getInt("total_downloads") + 1) {
+							PreparedStatement preparedStatement = connection
+									.prepareStatement("DELETE FROM smartshare.files WHERE id = ?;");
+							preparedStatement.setInt(1, resultSet.getInt("id"));
+							preparedStatement.execute();
+						} else { // If max downloads not reached
+							// Increment total_downloads
+							PreparedStatement preparedStatement = connection
+									.prepareStatement("UPDATE smartshare.files SET total_downloads = ? WHERE id = ?;");
+							preparedStatement.setInt(1, (resultSet.getInt("total_downloads") + 1));
+							preparedStatement.setInt(2, resultSet.getInt("id"));
+							preparedStatement.execute();
+						}
 					}
 
 				} else { // If it has expired delete entry
@@ -78,9 +84,15 @@ public class SQLRequestHandler {
 		return downloadedFile;
 	}
 
+	/**
+	 * Upload File into SQL Database
+	 *
+	 * @param uploadRequestDto JAXB annotated class representing the upload request
+	 * @return return true if uploaded, false if not uploaded
+	 */
 	public boolean setFile(UploadRequestDto uploadRequest) {
-		try (Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);) {
 
+		try (Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);) {
 			// Check to make sure there isn't already a file with the same name
 			// If there is return false
 			String sql = "SELECT * FROM smartshare.files WHERE file_name = ?;";
@@ -126,22 +138,41 @@ public class SQLRequestHandler {
 		return false;
 	}
 
+	/**
+	 * Get summary from SQL Database
+	 *
+	 * @param downloadRequestDto JAXB annotated class representing the summary
+	 *                           request
+	 * @return return SummaryDto, return null if not found or expired
+	 */
 	public SummaryDto getSummary(DownloadRequestDto downloadRequest) {
 		SummaryDto summary = null;
 		try (Connection connection = DriverManager.getConnection(URL, USERNAME, PASSWORD);) {
-			String sqlReturn = "SELECT * FROM smartshare.files WHERE file_name = ? AND password = ? ;";
-			PreparedStatement preparedStatementReturn = connection.prepareStatement(sqlReturn);
+			String sql = "SELECT * FROM smartshare.files WHERE file_name = ? AND password = ? ;";
+			PreparedStatement preparedStatementReturn = connection.prepareStatement(sql);
 			preparedStatementReturn.setString(1, downloadRequest.getFilename());
 			preparedStatementReturn.setString(2, downloadRequest.getPassword());
 			ResultSet resultSet = preparedStatementReturn.executeQuery();
 
 			while (resultSet.next()) {
-				// Fill summary information
-				summary = new SummaryDto();
-				
-				summary.setTimeCreated(resultSet.getTimestamp("time_created"));
-				summary.setDownloadsRemaining(resultSet.getInt("max_downloads") - resultSet.getInt("total_downloads"));
-				summary.setTimeTilExpiration(((int) (new Timestamp(resultSet.getTimestamp("expiry_time").getTime() - System.currentTimeMillis()).getTime())) / 60000);
+				// if not expired
+				if ((new Timestamp(System.currentTimeMillis())).before(resultSet.getTimestamp("expiry_time"))) {
+					// Fill summary information
+					summary = new SummaryDto();
+
+					summary.setTimeCreated(resultSet.getTimestamp("time_created"));
+					summary.setDownloadsRemaining(
+							resultSet.getInt("max_downloads") - resultSet.getInt("total_downloads"));
+					// Convert to minutes
+					summary.setTimeTilExpiration(((int) (new Timestamp(
+							resultSet.getTimestamp("expiry_time").getTime() - System.currentTimeMillis()).getTime()))
+							/ 60000);
+				} else { // If it has expired delete entry
+					sql = "DELETE FROM smartshare.files WHERE id = ?;";
+					PreparedStatement preparedStatement = connection.prepareStatement(sql);
+					preparedStatement.setInt(1, resultSet.getInt("id"));
+					preparedStatement.execute();
+				}
 			}
 		} catch (SQLException e) {
 			System.out.println("getSummary Failed: ");
